move smart_list into sub-package/multiple files
Step one of refactoring - making SmartList into its own
package, with each class having its own file.  No code
changes were made.

Note that SmartList and ListProxy import each other,
so had to import SmartList as a full package name
rather than use from ... import ... construct.
