+++ mwparserfromhell/smart_list/__init__.py
@@ -1,0 +1,1 @@
+from .SmartList import SmartList
